refactor(fix): move authOptions to seperate file
